Revert "Revert "Går direkt mot dokarkiv for knyttTilAnnensak (#150)" (#151)"
This reverts commit ab1ab35759d3e6a2c5c43c094e5ba8182ff6c93c.
