+++ nais/dev-gcp.json
@@ -5,7 +5,6 @@
   ],
   "externalHosts": [
     "dokarkiv.dev-fss-pub.nais.io",
-    "omsorgspenger-proxy.dev-fss.nais.io",
     "oppgave-q1.dev-fss-pub.nais.io",
     "saf.dev-fss-pub.nais.io"
   ],
@@ -16,8 +15,6 @@
     "DOKARKIV_SCOPES": "api://dev-fss.teamdokumenthandtering.dokarkiv-q1/.default",
     "OPPGAVE_BASE_URL": "https://oppgave-q1.dev-fss-pub.nais.io",
     "OPPGAVE_SCOPES": "api://dev-fss.oppgavehandtering.oppgave-q1/.default",
-    "DOKARKIVPROXY_BASE_URL": "https://omsorgspenger-proxy.dev-fss-pub.nais.io/dokarkivproxy",
-    "DOKARKIVPROXY_SCOPES": "824525c6-74df-47fc-8892-444e5f8506de/.default",
     "SAF_BASE_URL": "https://saf.dev-fss-pub.nais.io",
     "SAF_SCOPES": "api://dev-fss.teamdokumenthandtering.saf-q1/.default"
   }
