+++ src/test/kotlin/no/nav/omsorgspenger/testutils/wiremock/DokarkivMock.kt
@@ -15,8 +15,10 @@ private const val journalpostPath = "$basePath/rest/journalpostapi/v1/journalpos
 
 private fun oppdaterJournalpostMapping(
     callIdPattern: StringValuePattern = AnythingPattern()
-) = WireMock.put(WireMock
-    .urlMatching(".*$journalpostPath.*"))
+) = WireMock.put(
+    WireMock
+        .urlMatching(".*$journalpostPath.*")
+)
     .withHeader("Authorization", RegexPattern("^Bearer .+$"))
     .withHeader("Content-Type", equalTo("application/json"))
     .withHeader("Nav-Consumer-Id", equalTo("omsorgspenger-journalforing"))
@@ -29,20 +31,24 @@ private fun oppdaterJournalpostMapping(
 
 private fun opprettJournalpostMapping(
     callIdPattern: StringValuePattern = AnythingPattern()
-) = WireMock.post(WireMock
-    .urlMatching(".*$journalpostPath.*"))
+) = WireMock.post(
+    WireMock
+        .urlMatching(".*$journalpostPath.*")
+)
     .withQueryParam("forsoekFerdigstill", equalTo("true"))
     .withHeader("Authorization", RegexPattern("^Bearer .+$"))
     .withHeader("Content-Type", equalTo("application/json"))
     .withHeader("Nav-Consumer-Id", equalTo("omsorgspenger-journalforing"))
     .withHeader("Nav-Callid", callIdPattern)
     .withRequestBody(matchingJsonPath("$.sak.fagsaksystem", WireMock.matching("K9|OMSORGSPENGER")))
-    // Tester formatet på denn JSON'en i NyJournalpostTest, sjekker derfor her bare en prop
+// Tester formatet på denn JSON'en i NyJournalpostTest, sjekker derfor her bare en prop
 
 private fun ferdigstillJournalpostMapping(
     callIdPattern: StringValuePattern = AnythingPattern()
-) = WireMock.patch(WireMock
-    .urlPathMatching(".*$journalpostPath.*"))
+) = WireMock.patch(
+    WireMock
+        .urlPathMatching(".*$journalpostPath.*")
+)
     .withHeader("Authorization", RegexPattern("^Bearer .+$"))
     .withHeader("Content-Type", equalTo("application/json"))
     .withHeader("Nav-Consumer-Id", equalTo("omsorgspenger-journalforing"))
@@ -50,21 +56,23 @@ private fun ferdigstillJournalpostMapping(
     .withRequestBody(matchingJsonPath("$.journalfoerendeEnhet", containing("9999")))
 
 private fun WireMockServer.stubOppdaterJournalpostOk() = also {
-    stubFor(oppdaterJournalpostMapping()
-        .willReturn(
-            WireMock.aResponse()
-                .withStatus(200)
-        )
+    stubFor(
+        oppdaterJournalpostMapping()
+            .willReturn(
+                WireMock.aResponse()
+                    .withStatus(200)
+            )
     )
 }
 
 private fun WireMockServer.stubOppdaterJournalpostUventetFeil() = also {
-    stubFor(oppdaterJournalpostMapping(callIdPattern = equalTo("400"))
-        .willReturn(
-            WireMock.aResponse()
-                .withStatus(400)
-                .withBody("Noe gikk helt galt")
-        )
+    stubFor(
+        oppdaterJournalpostMapping(callIdPattern = equalTo("400"))
+            .willReturn(
+                WireMock.aResponse()
+                    .withStatus(400)
+                    .withBody("Noe gikk helt galt")
+            )
     )
 }
 
@@ -79,71 +87,104 @@ private fun WireMockServer.stubOppdaterJournalpostAlleredeFerdigstilt() = also {
         	"path": "/rest/journalpostapi/v1/journalpost/111111111"
         }
     """.trimIndent()
-    stubFor(oppdaterJournalpostMapping(callIdPattern = equalTo("allerede-ferdigstilt"))
-        .willReturn(
-            WireMock.aResponse()
-                .withStatus(400)
-                .withBody(json)
-        )
+    stubFor(
+        oppdaterJournalpostMapping(callIdPattern = equalTo("allerede-ferdigstilt"))
+            .willReturn(
+                WireMock.aResponse()
+                    .withStatus(400)
+                    .withBody(json)
+            )
     )
 }
 
 private fun WireMockServer.stubFerdigstillJournalpostOk() = also {
-    stubFor(ferdigstillJournalpostMapping()
-        .willReturn(
-            WireMock.aResponse()
-                .withStatus(200)
-        )
+    stubFor(
+        ferdigstillJournalpostMapping()
+            .willReturn(
+                WireMock.aResponse()
+                    .withStatus(200)
+            )
     )
 }
 
 private fun WireMockServer.stubFerdigstillJournalpostUventetFeil() = also {
-    stubFor(ferdigstillJournalpostMapping(callIdPattern = equalTo("feil-ved-ferdigstilling"))
-        .willReturn(
-            WireMock.aResponse()
-                .withStatus(400)
-        )
+    stubFor(
+        ferdigstillJournalpostMapping(callIdPattern = equalTo("feil-ved-ferdigstilling"))
+            .willReturn(
+                WireMock.aResponse()
+                    .withStatus(400)
+            )
     )
 }
 
 private fun WireMockServer.stubOpprettJournalpostOk() = also {
-    stubFor(opprettJournalpostMapping().willReturn(WireMock.aResponse()
-        .withStatus(201)
-        .withBody("""{"journalpostId":"12345678", "journalpostferdigstilt": true}""")
-    ))
+    stubFor(
+        opprettJournalpostMapping().willReturn(
+            WireMock.aResponse()
+                .withStatus(201)
+                .withBody("""{"journalpostId":"12345678", "journalpostferdigstilt": true}""")
+        )
+    )
 }
 
 private fun WireMockServer.stubOpprettJournalpostAlleredeOpprettet() = also {
-    stubFor(opprettJournalpostMapping(callIdPattern = equalTo("allerede-opprettet")).willReturn(WireMock.aResponse()
-        .withStatus(409)
-        .withBody("""{"journalpostId":"910111213", "journalpostferdigstilt": true}""")
-    ))
+    stubFor(
+        opprettJournalpostMapping(callIdPattern = equalTo("allerede-opprettet")).willReturn(
+            WireMock.aResponse()
+                .withStatus(409)
+                .withBody("""{"journalpostId":"910111213", "journalpostferdigstilt": true}""")
+        )
+    )
 }
 
 private fun WireMockServer.stubOpprettJournalpostIkkeFerdigstilt() = also {
-    stubFor(opprettJournalpostMapping(callIdPattern = equalTo("ikke-ferdigstilt")).willReturn(WireMock.aResponse()
-        .withStatus(201)
-        .withBody("""{"journalpostId":"1415161718", "journalpostferdigstilt": false}""")
-    ))
+    stubFor(
+        opprettJournalpostMapping(callIdPattern = equalTo("ikke-ferdigstilt")).willReturn(
+            WireMock.aResponse()
+                .withStatus(201)
+                .withBody("""{"journalpostId":"1415161718", "journalpostferdigstilt": false}""")
+        )
+    )
 }
 
 private fun WireMockServer.stubIsReady() = also {
-    stubFor(WireMock.get("$basePath/actuator/health/readiness")
-        .willReturn(WireMock.aResponse()
-            .withStatus(200)
-        )
+    stubFor(
+        WireMock.get("$basePath/actuator/health/readiness")
+            .willReturn(
+                WireMock.aResponse()
+                    .withStatus(200)
+            )
+    )
+}
+
+private fun WireMockServer.stubKnyttTilAnnenSakK9() = also { wireMockServer ->
+    wireMockServer.stubFor(
+        WireMock.put(WireMock.urlMatching(".*$journalpostPath.*"))
+            .withHeader("Authorization", RegexPattern("^Bearer .+$"))
+            .withHeader("Content-Type", WireMock.equalTo("application/json"))
+            .withHeader("Accept", WireMock.equalTo("application/json"))
+            .withHeader("Nav-Consumer-Id", WireMock.equalTo("omsorgspenger-journalforing"))
+            .withHeader("Nav-Callid", AnythingPattern())
+            .withRequestBody(WireMock.matchingJsonPath("$.sakstype", WireMock.equalTo("FAGSAK")))
+            .withRequestBody(WireMock.matchingJsonPath("$.fagsaksystem", WireMock.equalTo("K9")))
+            .withRequestBody(WireMock.matchingJsonPath("$.tema", WireMock.equalTo("OMS")))
+            .withRequestBody(WireMock.matchingJsonPath("$.journalfoerendeEnhet", WireMock.equalTo("9999")))
+            .withRequestBody(WireMock.matchingJsonPath("$.bruker.idType", WireMock.equalTo("FNR")))
+            .withRequestBody(WireMock.matchingJsonPath("$.bruker.id"))
+            .willReturn(WireMock.aResponse().withStatus(200).withBody("""{"nyJournalpostId":"123412341234"}"""))
     )
 }
 
 internal fun WireMockServer.stubDokarkiv() =
     stubOppdaterJournalpostOk()
-    .stubOppdaterJournalpostUventetFeil()
-    .stubOppdaterJournalpostAlleredeFerdigstilt()
-    .stubFerdigstillJournalpostOk()
-    .stubFerdigstillJournalpostUventetFeil()
-    .stubOpprettJournalpostOk()
-    .stubOpprettJournalpostAlleredeOpprettet()
-    .stubOpprettJournalpostIkkeFerdigstilt()
-    .stubIsReady()
+        .stubOppdaterJournalpostUventetFeil()
+        .stubOppdaterJournalpostAlleredeFerdigstilt()
+        .stubFerdigstillJournalpostOk()
+        .stubFerdigstillJournalpostUventetFeil()
+        .stubOpprettJournalpostOk()
+        .stubOpprettJournalpostAlleredeOpprettet()
+        .stubOpprettJournalpostIkkeFerdigstilt()
+        .stubIsReady()
+        .stubKnyttTilAnnenSakK9()
 
 internal fun WireMockServer.dokarkivBaseUrl() = baseUrl() + basePath
